improved chart
