+++ webapps/graph-chart/webapp.json
@@ -16,7 +16,7 @@
                 "name": "max_nodes",
                 "type": "INT",
                 "label": "Max displayed nodes",
-                "description": "",
+                "description": "+/- one node",
                 "mandatory": true,
                 "minI":1,
                 "maxI":2000,
